requirements.txt
pygame-ce
pyperclip